+++ xskillscore/__init__.py
@@ -2,6 +2,7 @@
 from pkg_resources import DistributionNotFound, get_distribution
 
 from .core.accessor import XSkillScoreAccessor
+from .core.comparative import sign_test
 from .core.contingency import Contingency
 from .core.deterministic import (
     effective_sample_size,